pyproject.toml
[project]
name = "youtube-multi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp>=2025.1.0",
    "scenedetect>=0.7",
    "opencv-python>=4.10",
    "pytesseract>=0.3.10",
    "pillow>=10.0",
]

[project.scripts]
multi = "youtube_multi.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_multi"]
